# Optimize REPL load