--- a/apps/repl/vite.config.js
+++ b/apps/repl/vite.config.js
@@ -22,6 +22,56 @@ export default defineConfig(() => ({
       'ember-repl',
       'repl-sdk',
     ],
+    // These dependencies are *always*
+    // needed on initial load.
+    // So we can boost initial load perf by eagerly optimizing them instead of waiting for the module graph crawl
+    include: [
+      // Our Runtime
+      '@shikijs/rehype/core',
+      'shiki/core',
+      'shiki/langs/*',
+      'shiki/wasm',
+      'shiki/themes/github-*',
+      // Framework
+      // 'ember-source/@ember/**/*',
+      // Theme and Syntax
+      '@codemirror/language',
+      '@codemirror/view',
+      // REPL + Editor
+      'ember-repl > codemirror',
+      'ember-repl > repl-sdk > codemirror',
+      'ember-repl > repl-sdk > codemirror-lang-mermaid',
+      'ember-repl > repl-sdk > tarparser',
+      'ember-repl > repl-sdk > package-name-regex',
+      'ember-repl > repl-sdk > rehype-raw',
+      'ember-repl > repl-sdk > rehype-stringify',
+      'ember-repl > repl-sdk > remark-gfm',
+      'ember-repl > repl-sdk > remark-parse',
+      'ember-repl > repl-sdk > remark-rehype',
+      'ember-repl > repl-sdk > unified',
+      'ember-repl > repl-sdk > unist-util-visit',
+      'ember-repl > repl-sdk > change-case',
+      'ember-repl > repl-sdk > @codemirror/autocomplete',
+      'ember-repl > repl-sdk > @codemirror/commands',
+      'ember-repl > repl-sdk > @codemirror/lang-html',
+      'ember-repl > repl-sdk > @codemirror/lang-html > @codemirror/lang-css',
+      'ember-repl > repl-sdk > @codemirror/lang-javascript',
+      'ember-repl > repl-sdk > @codemirror/lang-javascript > @lezer/javascript',
+      'ember-repl > repl-sdk > @codemirror/lang-markdown',
+      'ember-repl > repl-sdk > @codemirror/lang-markdown > @lezer/markdown',
+      'ember-repl > repl-sdk > @codemirror/lang-vue',
+      'ember-repl > repl-sdk > @codemirror/lang-yaml',
+      'ember-repl > repl-sdk > @codemirror/language',
+      'ember-repl > repl-sdk > @codemirror/language > @lezer/common',
+      'ember-repl > repl-sdk > @codemirror/language > @lezer/lr',
+      'ember-repl > repl-sdk > @codemirror/language > @lezer/highlight',
+      'ember-repl > repl-sdk > @codemirror/language-data',
+      'ember-repl > repl-sdk > @codemirror/lint',
+      'ember-repl > repl-sdk > @codemirror/search',
+      'ember-repl > repl-sdk > @codemirror/state',
+      'ember-repl > repl-sdk > @codemirror/view',
+    ],
+
     // for top-level-await, etc
     esbuildOptions: {
       target: 'esnext',