--- a/packages/repl-sdk/package.json
+++ b/packages/repl-sdk/package.json
@@ -50,7 +50,6 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@babel/plugin-proposal-decorators": "^7.28.0",
     "@nullvoxpopuli/eslint-configs": "^5.2.0",
     "@shikijs/rehype": "^3.7.0",
     "@tsconfig/ember": "^3.0.7",