Adds perf-check.sh, and I have these results atm:
```
| Dev Server Ready | First Paint | App Loaded |
| ---------------- | ----------- | ---------- |
| 842 ms           | 5,766 ms    | 7,879 ms   |
```

This is good because when I was on embroider-webpack initial load / time to first paint was closer to 25/30s.

Before customizing the optimizeDeps config, these were my results
```
| Dev Server Ready | First Paint | App Loaded |
| ---------------- | ----------- | ---------- |
| 989 ms           | 14,481 ms   | 32,307 ms  |
```

The subsequent reloads due to lazy depOptimization for the default path were silly.